backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1

# Document extraction
pypdf==5.4.0
python-docx==1.1.2
aiofiles==24.1.0

# Async SQLite for ingestion job + chunk tracking
aiosqlite==0.20.0

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
